+++ scp/lib/ui/cards.dart
@@ -2,8 +2,8 @@ import 'package:firebase_database/firebase_database.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:scp/booking.dart';
-import 'package:scp/mentor_search/mentee_page.dart';
-import 'package:scp/mentor_search/mentor_page.dart';
+import 'package:scp/ui/views/mentor_search/mentee_page.dart';
+import 'package:scp/ui/views/mentor_search/mentor_page.dart';
 import 'package:scp/ui/gradients.dart';
 import 'package:scp/firebase/firebaseDBHandler.dart';
 import 'package:scp/utils/routes.dart';
